requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
black==21.7b0
cfgv==3.3.0
click==8.0.1
coverage==5.5
distlib==0.3.2
Django==3.2.6
django-extensions==3.1.3
factory-boy==3.2.0
Faker==8.12.0
filelock==3.0.12
flake8==3.9.2
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
identify==2.2.13
inflection==0.5.1
iniconfig==1.1.1
isort==5.9.3
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.0
pathspec==0.9.0
platformdirs==2.2.0
pluggy==0.13.1
pre-commit==2.14.0
promise==2.3
py==1.10.0
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
snowballstemmer==2.1.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
virtualenv==20.7.2
